pyproject.toml
[project]
name = "genai-field-guide"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.7"
dependencies = [
    "google-adk>=1.18.0",
]
